Add hex and elf tasks
This is easier to remember across projects. This also solved the problem of the image task always executing its contents. We needed to make it a file task. But then you have to remember the directory and file name to call that file task, which isn't friendly. So instead these new simple task names were created.